back out module rename from django_irods to django_s3
