+++ hs_file_types/tests/test_view_functions.py
@@ -7,7 +7,7 @@ from django.test import RequestFactory, TestCase
 from django.urls import reverse
 from rest_framework import status
 
-from django_s3.views import download
+from django_irods.views import download
 from hs_core import hydroshare
 from hs_core.models import ResourceFile
 from hs_core.testing import MockS3TestCaseMixin
@@ -1840,7 +1840,7 @@ class TestFileTypeViewFunctions(MockS3TestCaseMixin, TestCase, CompositeResource
         # compute meta xml file storage path
         meta_file_storage_path = res_file.storage_path[:-4] + '_meta.xml'
         url_params = {'path': meta_file_storage_path}
-        url = reverse('django_s3_download', kwargs=url_params)
+        url = reverse('django_irods_download', kwargs=url_params)
         url = f"{url}?zipped=False&aggregation=False"
         request = self.factory.get(url)
         request.user = self.user
@@ -1854,7 +1854,7 @@ class TestFileTypeViewFunctions(MockS3TestCaseMixin, TestCase, CompositeResource
         # compute resmap  storage path
         meta_file_storage_path = res_file.storage_path[:-4] + '_resmap.xml'
         url_params = {'path': meta_file_storage_path}
-        url = reverse('django_s3_download', kwargs=url_params)
+        url = reverse('django_irods_download', kwargs=url_params)
         url = f"{url}?zipped=False&aggregation=False"
         request = self.factory.get(url)
         request.user = self.user
@@ -1889,7 +1889,7 @@ class TestFileTypeViewFunctions(MockS3TestCaseMixin, TestCase, CompositeResource
         # compute schema file storage path
         meta_file_storage_path = res_file.storage_path[:-4] + '_schema.json'
         url_params = {'path': meta_file_storage_path}
-        url = reverse('django_s3_download', kwargs=url_params)
+        url = reverse('django_irods_download', kwargs=url_params)
         url = f"{url}?zipped=False&aggregation=False"
         request = self.factory.get(url)
         request.user = self.user
@@ -1919,7 +1919,7 @@ class TestFileTypeViewFunctions(MockS3TestCaseMixin, TestCase, CompositeResource
         meta_file_storage_path = os.path.join(self.composite_resource.file_path, new_folder)
         meta_file_storage_path = os.path.join(meta_file_storage_path, f"{new_folder}_meta.xml")
         url_params = {'path': meta_file_storage_path}
-        url = reverse('django_s3_download', kwargs=url_params)
+        url = reverse('django_irods_download', kwargs=url_params)
         url = f"{url}?zipped=False&aggregation=False"
         request = self.factory.get(url)
         request.user = self.user
@@ -1934,7 +1934,7 @@ class TestFileTypeViewFunctions(MockS3TestCaseMixin, TestCase, CompositeResource
         meta_file_storage_path = os.path.join(self.composite_resource.file_path, new_folder)
         meta_file_storage_path = os.path.join(meta_file_storage_path, f"{new_folder}_resmap.xml")
         url_params = {'path': meta_file_storage_path}
-        url = reverse('django_s3_download', kwargs=url_params)
+        url = reverse('django_irods_download', kwargs=url_params)
         url = f"{url}?zipped=False&aggregation=False"
         request = self.factory.get(url)
         request.user = self.user
