+++ hydroshare/settings.py
@@ -408,7 +408,7 @@ INSTALLED_APPS = (
     "django.contrib.postgres",
     "django.contrib.messages",
     "django_nose",
-    "django_s3",
+    "django_irods",
     "drf_yasg",
     "theme",
     "theme.blog_mods",
@@ -488,7 +488,7 @@ APPS_TO_NOT_RUN = (
     "rest_framework",
     "django_nose",
     "grappelli_safe",
-    "django_s3",
+    "django_irods",
     "crispy_forms",
     "autocomplete_light",
     "widget_tweaks",
