+++ local-dev-first-start-only.sh
@@ -310,11 +310,6 @@ echo " Migrating data"
 echo '########################################################################################################################'
 echo
 
-echo
-echo "  - docker exec -u hydro-service hydroshare python manage.py rename_app django_irods django_s3"
-echo
-docker exec -u hydro-service hydroshare python manage.py rename_app django_irods django_s3
-
 echo "  - docker exec hydroshare chown -R hydro-service:storage-hydro /tmp /shared_tmp"
 docker exec hydroshare chown -R hydro-service:storage-hydro /tmp /shared_tmp
 echo
